🌱 [scanner] refactor: further split lib/api.ts into focused modules
Build on the initial types/endpoints extraction by moving the remaining
~940 lines of api.ts into four focused modules under api/:

- api/helpers.ts  (~77 lines): response parsing utilities
  (safeJson, safeParseJsonOrNull, safeReadTextOrEmpty, isAbortError,
  createErrorWithCause)
- api/session.ts  (~209 lines): session/auth error handling
  (handle401, handle429, performSessionExpiry, showSessionExpiredBanner)
- api/backend.ts  (~262 lines): backend availability tracking and
  OAuth probing (checkBackendAvailability, markBackend{Failure,Success},
  isBackendUnavailable, checkOAuthConfigured, checkOAuthConfiguredWithRetry)
- api/client.ts   (~463 lines): ApiClient class (get/post/patch/put/delete),
  api singleton, and authFetch helper

api.ts is now a 40-line re-export barrel; all public exports remain stable
so no callers are changed.
